+++ Editor/Brushes/LineBrush/LineBrush.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Unity.Collections;
 using UnityEngine;
 using UnityEngine.Tilemaps;
 
@@ -54,17 +55,22 @@ namespace UnityEditor.Tilemaps
         {
             if (lineStartActive)
             {
+                if (Event.current != null &&
+                    Event.current.type == EventType.MouseDrag)
+                    return;
+
                 var startPos = new Vector2Int(lineStart.x, lineStart.y);
                 var endPos = new Vector2Int(position.x, position.y);
                 if (startPos == endPos)
                     base.Paint(grid, brushTarget, position);
                 else
+                {
                     foreach (var point in GetPointsOnLine(startPos, endPos, fillGaps))
                     {
                         var paintPos = new Vector3Int(point.x, point.y, position.z);
                         base.Paint(grid, brushTarget, paintPos);
                     }
-
+                }
                 lineStartActive = false;
             }
             else if (IsMoving)
@@ -239,6 +245,8 @@ namespace UnityEditor.Tilemaps
         private Texture2D m_BrushIcon;
         private LineBrush lineBrush => target as LineBrush;
 
+        private Vector3[] startCells = new Vector3[4];
+
         /// <summary> Returns an icon identifying the Line Brush. </summary>
         public override Texture2D icon
         {
@@ -255,8 +263,31 @@ namespace UnityEditor.Tilemaps
         }
 
         /// <summary>
+        ///     Callback for handling the GUI events for the GridBrush in the Scene View.
+        ///     The LineBrush Editor overrides this cancel drawing lines with the right mouse button.
+        /// </summary>
+        /// <param name="gridLayout">Grid that the brush is being used on.</param>
+        /// <param name="brushTarget">Target of the GridBrushBase::ref::Tool operation.
+        ///     By default the currently selected GameObject.</param>
+        public override void OnSceneGUI(GridLayout gridLayout, GameObject brushTarget)
+        {
+            base.OnSceneGUI(gridLayout, brushTarget);
+
+            if (!lineBrush.lineStartActive)
+                return;
+
+            if (Event.current.type == EventType.MouseDown
+                && Event.current.button == 1)
+            {
+                lineBrush.lineStartActive = false;
+                Event.current.Use();
+                return;
+            }
+        }
+
+        /// <summary>
         ///     Callback for painting the GUI for the GridBrush in the Scene View.
-        ///     The CoordinateBrush Editor overrides this to draw the preview of the brush when drawing lines.
+        ///     The LineBrush Editor overrides this to draw the preview of the brush when drawing lines.
         /// </summary>
         /// <param name="grid">Grid that the brush is being used on.</param>
         /// <param name="brushTarget">
@@ -270,7 +301,7 @@ namespace UnityEditor.Tilemaps
             GridBrushBase.Tool tool, bool executing)
         {
             base.OnPaintSceneGUI(grid, brushTarget, position, tool, executing);
-            if (lineBrush.lineStartActive && brushTarget != null)
+            if (brushTarget != null)
             {
                 var tilemap = brushTarget.GetComponent<Tilemap>();
                 if (tilemap != null)
@@ -279,36 +310,122 @@ namespace UnityEditor.Tilemaps
                     lastTilemap = tilemap;
                 }
 
+                if (!lineBrush.lineStartActive)
+                    return;
+
                 // Draw preview tiles for tilemap
                 var startPos = new Vector2Int(lineBrush.lineStart.x, lineBrush.lineStart.y);
                 var endPos = new Vector2Int(position.x, position.y);
                 if (startPos == endPos)
                     PaintPreview(grid, brushTarget, position.min);
                 else
+                {
                     foreach (var point in LineBrush.GetPointsOnLine(startPos, endPos, lineBrush.fillGaps))
                     {
                         var paintPos = new Vector3Int(point.x, point.y, position.z);
                         PaintPreview(grid, brushTarget, paintPos);
                     }
+                }
 
                 if (Event.current.type == EventType.Repaint)
                 {
-                    var min = lineBrush.lineStart;
-                    var max = lineBrush.lineStart + position.size;
-
-                    // Draws a box on the picked starting position
                     GL.PushMatrix();
-                    GL.MultMatrix(GUI.matrix);
+                    GL.MultMatrix(tilemap.transform.localToWorldMatrix);
                     GL.Begin(GL.LINES);
-                    Handles.color = Color.blue;
-                    Handles.DrawLine(new Vector3(min.x, min.y, min.z), new Vector3(max.x, min.y, min.z));
-                    Handles.DrawLine(new Vector3(max.x, min.y, min.z), new Vector3(max.x, max.y, min.z));
-                    Handles.DrawLine(new Vector3(max.x, max.y, min.z), new Vector3(min.x, max.y, min.z));
-                    Handles.DrawLine(new Vector3(min.x, max.y, min.z), new Vector3(min.x, min.y, min.z));
+                    GL.Color(Color.blue);
+
+                    var startArea = new BoundsInt(lineBrush.lineStart, position.size);
+                    if (grid.cellLayout == GridLayout.CellLayout.Hexagon)
+                    {
+                        var cellLocals = GetHexOutlineVertices(grid, startArea);
+                        if (cellLocals.IsCreated)
+                        {
+                            for (int i = 0, j = cellLocals.Length - 1; i < cellLocals.Length; j = i++)
+                            {
+                                GL.Vertex3(cellLocals[j].x, cellLocals[j].y, cellLocals[j].z);
+                                GL.Vertex3(cellLocals[i].x, cellLocals[i].y, cellLocals[i].z);
+                            }
+                            cellLocals.Dispose();
+                        }
+                    }
+                    else
+                    {
+                        var min = lineBrush.lineStart;
+                        var max = lineBrush.lineStart + position.size;
+
+                        var cellStride = tilemap.cellSize + tilemap.cellGap;
+                        var cellGap = Vector3.one;
+                        if (!Mathf.Approximately(cellStride.x, 0f))
+                        {
+                            cellGap.x = tilemap.cellSize.x / cellStride.x;
+                        }
+                        if (!Mathf.Approximately(cellStride.y, 0f))
+                        {
+                            cellGap.y = tilemap.cellSize.y / cellStride.y;
+                        }
+
+                        startCells[0] = tilemap.CellToLocalInterpolated(new Vector3(min.x, min.y, min.z));
+                        startCells[1] = tilemap.CellToLocalInterpolated(new Vector3(max.x - 1 + cellGap.x, min.y, min.z));
+                        startCells[2] = tilemap.CellToLocalInterpolated(new Vector3(max.x - 1 + cellGap.x, max.y - 1 + cellGap.y, min.z));
+                        startCells[3] = tilemap.CellToLocalInterpolated(new Vector3(min.x, max.y - 1 + cellGap.y, min.z));
+
+                        // Draws a box on the picked starting position
+                        int i = 0;
+                        for (int j = startCells.Length - 1; i < startCells.Length; j = i++)
+                        {
+                            GL.Vertex(startCells[j]);
+                            GL.Vertex(startCells[i]);
+                        }
+                    }
                     GL.End();
                     GL.PopMatrix();
                 }
             }
+        }
+
+        internal static NativeArray<Vector3> GetHexOutlineVertices(GridLayout grid, BoundsInt area)
+        {
+            int requiredVertices = 4 * (area.size.x + area.size.y) - 2;
+            if (requiredVertices <= 0)
+                return default;
+
+            var cellLocals = new NativeArray<Vector3>(requiredVertices, Allocator.Temp);
+            int horizontalCount = area.size.x * 2;
+            int verticalCount = area.size.y * 2 - 1;
+            int bottom = 0;
+            int top = horizontalCount + verticalCount + horizontalCount - 1;
+            int left = requiredVertices - 1;
+            int right = horizontalCount;
+
+            var cellOffset = new NativeArray<Vector3>(6, Allocator.Temp);
+            cellOffset[0] = Grid.Swizzle(grid.cellSwizzle, new Vector3(0, grid.cellSize.y / 2, 0));
+            cellOffset[1] = Grid.Swizzle(grid.cellSwizzle, new Vector3(grid.cellSize.x / 2, grid.cellSize.y / 4, 0));
+            cellOffset[2] = Grid.Swizzle(grid.cellSwizzle, new Vector3(grid.cellSize.x / 2, -grid.cellSize.y / 4, 0));
+            cellOffset[3] = Grid.Swizzle(grid.cellSwizzle, new Vector3(0, -grid.cellSize.y / 2, 0));
+            cellOffset[4] = Grid.Swizzle(grid.cellSwizzle, new Vector3(-grid.cellSize.x / 2, -grid.cellSize.y / 4, 0));
+            cellOffset[5] = Grid.Swizzle(grid.cellSwizzle, new Vector3(-grid.cellSize.x / 2, grid.cellSize.y / 4, 0));
+
+            for (int x = area.min.x; x < area.max.x; x++)
+            {
+                cellLocals[bottom++] = grid.CellToLocal(new Vector3Int(x, area.min.y, area.zMin)) + cellOffset[4];
+                cellLocals[bottom++] = grid.CellToLocal(new Vector3Int(x, area.min.y, area.zMin)) + cellOffset[3];
+                cellLocals[top--] = grid.CellToLocal(new Vector3Int(x, area.max.y - 1, area.zMin)) + cellOffset[0];
+                cellLocals[top--] = grid.CellToLocal(new Vector3Int(x, area.max.y - 1, area.zMin)) + cellOffset[1];
+            }
+            cellLocals[left--] = grid.CellToLocal(new Vector3Int(area.min.x, area.min.y, area.zMin)) + cellOffset[5];
+            cellLocals[top--] = grid.CellToLocal(new Vector3Int(area.max.x - 1, area.max.y - 1, area.zMin)) + cellOffset[2];
+            for (int y = area.min.y + 1; y < area.max.y; y++)
+            {
+                cellLocals[left--] = grid.CellToLocal(new Vector3Int(area.min.x, y, area.zMin)) + cellOffset[4];
+                cellLocals[left--] = grid.CellToLocal(new Vector3Int(area.min.x, y, area.zMin)) + cellOffset[5];
+            }
+            for (int y = area.min.y; y < area.max.y - 1; y++)
+            {
+                cellLocals[right++] = grid.CellToLocal(new Vector3Int(area.max.x - 1, y, area.zMin)) + cellOffset[2];
+                cellLocals[right++] = grid.CellToLocal(new Vector3Int(area.max.x - 1, y, area.zMin)) + cellOffset[1];
+            }
+            cellOffset.Dispose();
+            return cellLocals;
         }
 
         /// <summary>